// src/interfaces/enums.ts
export enum UserRole {
  EDITOR = 'editor',
  VIEWER = 'viewer',
}

export enum ResourceType {
  ROOT = 'root',
  CHILD = 'child',
}

export enum LockStatus {
  LOCKED = 'locked',
  UNLOCKED = 'unlocked',
}

export enum PresenceStatus {
  ONLINE = 'online',
  AWAY = 'away',
  OFFLINE = 'offline',
}